Move aws.ReadSeekCloser and aws.WriteAtBuffer to feature/s3/manager.
