+++ feature/s3/manager/types_test.go
@@ -1,4 +1,4 @@
-package aws
+package manager
 
 import (
 	"bytes"
